requirements.txt
absl-py==1.4.0
accelerate==0.30.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.8.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
Automat==0.8.0
Babel==2.10.3
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
batchspawner==1.2.0
beautifulsoup4==4.12.2
bert-score==0.3.11
bitsandbytes==0.42.0
bleach==5.0.1
blessed==1.20.0
blinker==1.4
blis==0.7.11
boto3==1.34.86
botocore==1.34.86
Brotli==1.0.9
cachetools==4.2.4
catalogue==2.0.10
certifi==2023.11.17
certipy==0.1.3
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==7.1.2
clip==1.0
cloud-init==21.4
cloudpathlib==0.18.1
cmake==3.26.3
colorama==0.4.3
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
configparser==5.3.0
constantly==15.1.0
contourpy==1.0.7
croniter==1.4.1
cryptography==40.0.2
cssselect==1.2.0
cupshelpers==1.0
cupy-cuda11x==12.3.0
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.6.4
datasets==2.19.1
dateutils==0.6.12
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
deepdiff==6.5.0
defer==1.0.6
defusedxml==0.7.1
deplacy==2.1.0
dill==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.8.0
emoji==2.12.1
en-core-web-sm==3.6.0
entrypoints==0.3
esupar==1.7.2
evaluate==0.4.1
exceptiongroup==1.1.3
executing==1.1.1
fastapi==0.103.1
fastjsonschema==2.16.2
fastrlock==0.8.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
fire==0.5.0
-e git+https://github.com/FlagOpen/FlagEmbedding.git@a6069c10740e1198806ca638516e27caa7cde4ad#egg=FlagEmbedding
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.4.0
fsspec==2023.12.2
ftfy==6.1.3
gast==0.4.0
gensim==3.8.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3.post0
grpcio==1.43.0
h11==0.14.0
h5py==3.8.0
hickle==5.0.3
httplib2==0.14.0
huggingface-hub==0.23.2
hyperlink==19.0.0
idna==3.6
importlib-metadata==5.0.0
importlib-resources==5.10.0
incremental==16.10.1
inquirer==3.1.3
install==1.3.5
ipykernel==6.16.2
ipython==8.5.0
ipython-genutils==0.2.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jax==0.4.10
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.10
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.4.3
jupyter-core==4.11.2
jupyter-server==1.21.0
jupyter-telemetry==0.1.0
jupyterhub==3.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.16.1
keras==2.12.0
keyring==18.0.1
kiwisolver==1.4.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.44
langchain-text-splitters==0.0.1
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.49
language-data==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.0
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
lit==16.0.5
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.2.3
marisa-trie==1.2.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.1.0
more-itertools==4.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.1
newspaper3k==0.2.8
ninja==1.11.1.1
nltk==3.8.1
notebook==6.5.1
notebook-shim==0.2.0
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
opencv-python==4.5.3.56
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.1
packaging==21.3
pamela==1.0.0
pandas==1.3.0
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
peft==0.11.1
perfect==0.0.1
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.2.0
plac==1.1.3
preshed==3.0.8
prettytable==3.10.2
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.31
Protego==0.2.1
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.17.2
pyarrow==12.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.2.1
pybcj==1.0.1
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.18.0
pycups==1.9.73
pydantic==1.7.4
pydantic-core==2.18.4
pydeck==0.8.0
PyDispatcher==2.0.7
Pygments==2.16.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.8.0
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyppmd==1.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pysmb==1.2.9.1
pyspnego==0.9.0
python-apt==2.0.0+ubuntu0.20.4.6
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-editor==1.0.4
python-json-logger==2.0.4
python-multipart==0.0.6
python-version==0.0.2
pytorch-lightning==2.0.9
pytrec-eval==0.5
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pyvi==0.1.1
PyYAML==6.0.1
pyzmq==24.0.1
pyzstd==0.15.7
queuelib==1.6.2
rank-bm25==0.2.2
ray==1.11.0
readchar==4.0.5
redis==4.5.5
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
responses==0.18.0
rich==13.5.3
rouge-score==0.1.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.10.1
sacremoses==0.1.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
Scrapy==2.9.0
seaborn==0.13.2
SecretStorage==2.3.1
Send2Trash==1.8.0
sentence-transformers==2.7.0
sentencepiece==0.1.99
sentry-sdk==1.24.0
seqeval==1.2.2
service-identity==18.1.0
sgmllib3k==1.0.0
shortuuid==1.0.11
shtab==1.6.5
simplejson==3.16.0
simpletransformers==0.64.3
six==1.16.0
sklearn-crfsuite==0.5.0
smart-open==6.3.0
smbprotocol==1.10.1
smmap==5.0.0
sniffio==1.3.0
sos==4.1
soupsieve==2.5
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.42
srsly==2.4.8
ssh-import-id==5.10
stack-data==0.5.1
stanza==1.8.2
starlette==0.27.0
starsessions==1.3.0
streamlit==1.26.0
subprocess32==3.5.4
supar==1.1.4
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.5
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.16.0
texttable==1.6.7
thinc==8.1.12
threadpoolctl==3.2.0
tiktoken==0.7.0
timm==1.0.7
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==3.4.4
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchmetrics==1.1.2
torchvision==0.17.2+cu121
tornado==6.2
tqdm==4.66.1
traitlets==5.10.0
transformers==4.38.2
transformers-stream-generator==0.0.5
triton==2.2.0
trl==0.8.6
Twisted==18.9.0
typer==0.3.2
typing-extensions==4.9.0
typing-inspect==0.9.0
tyro==0.6.0
tzdata==2023.4
tzlocal==4.3.1
ubuntu-advantage-tools==27.5
ufw==0.36
ujson==4.0.2
unattended-upgrades==0.1
urllib3==2.1.0
uvicorn==0.23.2
validators==0.22.0
vi-core-news-lg==3.6.0
w3lib==2.1.1
wadllib==1.3.3
wandb==0.12.9
wasabi==0.10.1
watchdog==3.0.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.4
Wikidata==0.7.0
wikipedia==1.4.0
wrapt==1.14.1
xxhash==3.2.0
yarl==1.9.2
yaspin==2.3.0
zipp==3.15.0
zope.interface==6.0
